Se corrige nombre archivo que sea congruente con nombre de clase
