+++ app/src/dao/EmployeeDAO.java
@@ -1,7 +1,24 @@
 package dao;
 
+import org.hibernate.HibernateException;
+
+
 import model.Employee;
 
 public class EmployeeDAO extends BaseDAO<Employee> {
-	
+
+	public Employee getWithManagedTickets(long id) throws HibernateException {
+	    Employee employee = null;
+	    try {
+	        initTransaction();
+	        employee = (Employee) session.createQuery(
+	            "select e from Employee e left join fetch e.managedTickets where e.id = :id"
+	        ).setParameter("id", id)
+	         .uniqueResult();
+	    } finally {
+	        session.close();
+	    }
+	    return employee;
+	}
 }
+
